feat: Добавлена проверка id перед удалением карточки

+++ controllers/cards.js
@@ -1,7 +1,7 @@
 const { StatusCodes } = require('http-status-codes');
 
 const {
-  CREATED, BAD_REQUEST, INTERNAL_SERVER_ERROR, NOT_FOUND, OK,
+  CREATED, BAD_REQUEST, INTERNAL_SERVER_ERROR, NOT_FOUND, OK, FORBIDDEN
 } = StatusCodes;
 const { default: mongoose } = require('mongoose');
 const Card = require('../models/card');
@@ -36,9 +36,18 @@ const getAllCards = (req, res) => {
 // Удаление карточки
 const removeCard = (req, res) => {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
-    .orFail()
-    .then((card) => res.status(OK).send(card))
+  // Перед удалением проверяю на соответствие id пользователя и создателя карточки
+  Card.findById(cardId)
+    .then((card) => {
+      if (!card.owner.equals(req.user._id)) {
+        res.status(FORBIDDEN).send({message: 'Ошибка доступа'});
+        return;
+      }
+      Card.findByIdAndRemove(cardId)
+      .orFail()
+      .then((card) => res.status(OK).send(card))
+      .catch(() => res.status(INTERNAL_SERVER_ERROR).send(defaultServerError))
+    })
     .catch((err) => {
       switch (err.constructor) {
         case mongoose.Error.DocumentNotFoundError:
